Created web app

+++ TestMasGlobal.CalculatedSalary.Domain.Entities/Employee.cs
@@ -7,7 +7,7 @@
         public int RoleId { get; set; }
         public string RoleName { get; set; }
         public string RoleDescription { get; set; }
-        public int HourlySalary { get; set; }
-        public int MonthlySalary { get; set; }
+        public decimal HourlySalary { get; set; }
+        public decimal MonthlySalary { get; set; }
     }
 }
